requirements.txt
aiofiles==0.6.0
aiosqlite==0.16.0
asyncpg==0.21.0
certifi==2020.11.8
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
iso8601==0.1.13
multidict==5.0.0
PyJWT==1.7.1
PyPika==0.44.0
pytz==2020.4
rfc3986==1.4.0
sanic==20.9.1
scrypt==0.8.17
sniffio==1.2.0
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
websockets==8.1
